Remove caching support.
This is not part of the core goal for `folderify`, and could be managed with a wrapper in the future.

More importantly, the `--no-trim` feature means that there is more than one way to turn a cached mask into an icon, so we'd need to add a bunch more code to make restoration more reliable.