Truncate progress messages if they don't in the terminal.
This fixes moderately narrow terminals. However, it still doesn't address terminals with fewer than 38 columns.
We could remove more data (prefix and/or progress fraction) if the terminal is narrow, but that might still run into issues with terminals that are resized after we get the width. The proper solution is probably to wait on a fix for https://github.com/console-rs/indicatif/issues/520