v4.0.0
Release notes:

There are no functionality changes. However, there is one breaking change: instead of the deprecated `convert` command for ImageMagick, `folderify` will invoke `magick`. The `magick` command was added at least ≈8 years ago, so this should not be an issue for most users.

This change should prevent `WARNING: The convert command is deprecated in IMv7, use "magick"` messages from spilling into the output of `folderify`, particularly for Homebrew users who are on the latest version of ImageMagick.